requirements.txt
py
packaging
PyInstaller
pytest_cov
pytest_pep8
sphinx
sphinx_bootstrap_theme
sphinxcontrib-autoprogram
